+++ src/application.h
@@ -23,7 +23,7 @@ class Application
       void printMenu() const noexcept;
       void clearInputStream() const;  
       void verifyUserInputMenu(std::string& input) const;
-      int receiveUserInputMenu() const;
+      uint8_t receiveUserInputMenu() const;
       void userInterface(bool& keepGoing);
 };
